Add data.sql file

+++ src/main/java/com/softserve/skillscope/talent/TalentRepository.java
@@ -3,5 +3,5 @@ package com.softserve.skillscope.talent;
 import com.softserve.skillscope.talent.model.entity.Talent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TalentRepository extends JpaRepository<Talent, Integer> {
+public interface TalentRepository extends JpaRepository<Talent, Long> {
 }
